extraigo funciones a iterator y a rhs_iterator
y acomodo imports
